+++ FE/src/app/useDebounce.ts
@@ -1,0 +1,18 @@
+"use client";
+import { useEffect, useState } from "react";
+
+export const useDebounce = (inputValue: string, delay: number) => {
+  const [debouncedValue, setDebouncedValue] = useState(inputValue);
+
+  useEffect(() => {
+    const handler = setTimeout(() => {
+      setDebouncedValue(inputValue);
+    }, delay);
+
+    return () => {
+      clearTimeout(handler);
+    };
+  }, [inputValue, delay]);
+
+  return debouncedValue;
+};
